% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_012gt.R
\name{index_012gt}
\alias{index_012gt}
\title{Index '012'-formatted genotype file}
\usage{
index_012gt(genotype_file, scaffold_file)
}
\arguments{
\item{genotype_file}{the output of make_012gt() or make_012gt_from_vcf()}

\item{scaffold_file}{a file with the same format as used in RecView ShinyApp.}
}
\description{
Index the genotype file for faster access in RecView ShinyApp.
}
\note{
The file will be saved with the extension of .idx
}
