% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_gt012.R
\name{index_gt012}
\alias{index_gt012}
\title{Index '012'-formatted genotype file}
\usage{
index_gt012(gt012_file, scaffold_file)
}
\arguments{
\item{gt012_file}{the output of make_gt012() or make_012gt_from_vcf()}

\item{scaffold_file}{a file with the same format as used in RecView ShinyApp.}
}
\description{
Index the genotype file for faster access in RecView ShinyApp.
}
\note{
The file will be saved with the extension of .idx
}
