% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_012gt.R
\name{make_012gt}
\alias{make_012gt}
\title{Make '012'-formatted genotype file}
\usage{
make_012gt(gt, F0F1_rename, save_filename)
}
\arguments{
\item{gt}{the output file when using vcftools to extract "GT"}

\item{F0F1_rename}{a vector of the labels of F0 (father's father, father's mother, mother's father, mother's mother) and F1 (father, mother); use "NA" if missing.}

\item{save_filename}{the name for saving the '012'-formatted output.}

\item{readable}{if TRUE, a CSV file will also be produced.}
}
\description{
Make the genotype file in designated '012' format for RecView ShinyApp.
}
\note{
If setting readable=TRUE, it needs to set the extension as ".csv" in save_filename.
}
