% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_algorithm.R
\name{PD_algorithm}
\alias{PD_algorithm}
\title{Proportional Difference (PD) algorithm}
\usage{
PD_algorithm(x, value_col = "...", window_size = 400, min_threshold = 0.5, full_result = FALSE)
}
\arguments{
\item{x}{a data frame or tibble.}

\item{value_col}{character; the column in which the change points will be analysed. This column should contain binary data, e.g. 0 and 1.}

\item{window_size}{numeric; the size of sliding windows in which the proportion of symbol is checked. Default: 400.}

\item{min_threshold}{numeric; minimal threshold to detect optimal threshold for locating the change points. Default: 0.5.}

\item{full_result}{logical; it indicates whether the full results will be produced.}
}
\description{
Proportional Difference (PD) algorithm for locating change points.
}
