% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RecView.R
\name{plot.RecView}
\alias{plot.RecView}
\title{Plotting function for S3 class "RecView"}
\usage{
plot(obj)
}
\arguments{
\item{obj}{S3 Class "RecView".}
}
\description{
Plotting function for S3 class "RecView".
}
