% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_RecView_App.R
\name{run_RecView_App}
\alias{run_RecView_App}
\title{Run RecView ShinyApp}
\usage{
run_RecView_App()
}
\description{
Initiate the RecView ShinyApp. The RecView ShinyApp is intended to visualize and provide options automatically locate recombinations.
}
\note{
Before execute run_RecView_App(), it is recommmended to navigate the working directory to the directory of input files. This makes it convenient for saving result output files from the app.
}
