% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_012gt.R
\name{make_012gt}
\alias{make_012gt}
\title{Make '012'-formatted genotype file}
\usage{
make_012gt(path = "...", F0F1_rename = c("PGF", "PGM", "MGF", "MGM", "Father", "Mother"), save_filename = "gt_012.csv", compact = FALSE)
}
\arguments{
\item{path}{the path to the file of using vcftools to extract "GT".}

\item{F0F1_rename}{a vector of the labels of F0 (father's father, father's mother, mother's father, mother's mother) and F1 (father, mother); use "NA" if missing.}

\item{save_filename}{the name for saving the '012'-formatted output.}

\item{compact}{if TRUE, a binary format RDS file will also be produced, which takes up smaller storage space.}
}
\description{
Make the genotype file in designated '012' format for RecView ShinyApp.
}
