% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_2gen.R
\name{rec_2gen}
\alias{rec_2gen}
\title{Locate recombination with two-generation pedigree}
\usage{
rec_2gen(data, sc_order, chromosome, offspring, precision)
}
\arguments{
\item{data}{'012'-formatted genotype file, which is the output from make_012gt() or make_012gt_from_vcf().}

\item{sc_order}{scaffold file.}

\item{chromosome}{chromosome.}

\item{offspring}{the offspring to be included in the analysis.}

\item{precision}{the maximal base pairs in which the detected recombination positions will be pooled.}

\item{method}{the algorithm to identify the change point. "CCS", "PD", or "changepoint".}
}
\description{
Locate recombination with two-generation pedigree that consists of father, mother and at least two offspring
}
\note{
This function produces a S3 class "RecView".
}
